+++ packages/wallet-ts/src/strategies/wallet-strategy/strategies/Base.ts
@@ -1,18 +1,12 @@
 import { ChainId, EthereumChainId } from '@injectivelabs/ts-types'
-import { WalletStrategyEthereumOptions } from '../../types'
+import { WalletStrategyArguments } from '../../types'
 
 export default abstract class BaseConcreteStrategy {
   protected chainId: ChainId
 
   protected ethereumChainId?: EthereumChainId
 
-  protected constructor({
-    ethereumOptions,
-    chainId,
-  }: {
-    chainId: ChainId
-    ethereumOptions?: WalletStrategyEthereumOptions
-  }) {
+  protected constructor({ chainId, ethereumOptions }: WalletStrategyArguments) {
     this.ethereumChainId = ethereumOptions
       ? ethereumOptions.ethereumChainId
       : undefined