+++ packages/wallet-ts/src/strategies/wallet-strategy/strategies/Leap.ts
@@ -157,9 +157,9 @@ export default class Leap
     )
   }
 
-  async getChainId(): Promise<string> {
+  async getEthereumChainId(): Promise<string> {
     throw new CosmosWalletException(
-      new Error('getChainId is not supported on Leap'),
+      new Error('getEthereumChainId is not supported on Leap'),
       {
         code: UnspecifiedErrorCode,
         context: WalletAction.GetChainId,