+++ packages/sdk-ts/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.10.37](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/sdk-ts@1.10.36...@injectivelabs/sdk-ts@1.10.37) (2023-04-10)
+
+**Note:** Version bump only for package @injectivelabs/sdk-ts
+
+
+
+
+
+## [1.10.36](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/sdk-ts@1.10.35...@injectivelabs/sdk-ts@1.10.36) (2023-04-10)
+
+**Note:** Version bump only for package @injectivelabs/sdk-ts
+
+
+
+
+
+## [1.10.35](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/sdk-ts@1.10.34...@injectivelabs/sdk-ts@1.10.35) (2023-04-09)
+
+**Note:** Version bump only for package @injectivelabs/sdk-ts
+
+
+
+
+
 ## [1.10.34](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/sdk-ts@1.10.33...@injectivelabs/sdk-ts@1.10.34) (2023-04-08)