+++ packages/wallet-ts/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.10.41](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/wallet-ts@1.10.40...@injectivelabs/wallet-ts@1.10.41) (2023-04-10)
+
+**Note:** Version bump only for package @injectivelabs/wallet-ts
+
+
+
+
+
+## [1.10.40](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/wallet-ts@1.10.39...@injectivelabs/wallet-ts@1.10.40) (2023-04-10)
+
+**Note:** Version bump only for package @injectivelabs/wallet-ts
+
+
+
+
+
+## [1.10.39](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/wallet-ts@1.10.38...@injectivelabs/wallet-ts@1.10.39) (2023-04-09)
+
+**Note:** Version bump only for package @injectivelabs/wallet-ts
+
+
+
+
+
 ## [1.10.38](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/wallet-ts@1.10.37...@injectivelabs/wallet-ts@1.10.38) (2023-04-08)
 
 **Note:** Version bump only for package @injectivelabs/wallet-ts